requirements.txt
# Core scientific computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0

# Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Deep learning
torch>=1.9.0,<2.0.0
torchvision>=0.10.0,<1.0.0

# Specialized scientific tools
pysr>=0.8.0,<1.0.0  # Symbolic regression
gudhi>=3.4.0,<4.0.0  # Topological data analysis
scikit-image>=0.18.0,<1.0.0  # Image processing
h5py>=3.0.0,<4.0.0  # HDF5 file support

# Utilities
pyyaml>=5.4.0,<7.0.0  # YAML configuration
tqdm>=4.62.0,<5.0.0  # Progress bars
click>=8.0.0,<9.0.0  # CLI tools
python-dotenv>=0.19.0,<1.0.0  # Environment variables

# Data validation
pydantic>=1.8.0,<2.0.0  # Data validation
jsonschema>=3.2.0,<5.0.0  # JSON schema validation